refactor(ui): clean up audit export components
